requirements.txt
Bottleneck==1.3.5
brotlipy==0.7.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.4
click==8.0.4
colorama==0.4.5
cryptography==37.0.1
Flask==2.0.3
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.1.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
numexpr==2.8.3
numpy==1.22.3
packaging==21.3
pandas==1.3.5
pip==21.2.4
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.4
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.0
scikit-learn==1.0.1
scipy==1.7.3
setuptools==61.2.0
six==1.16.0
spotipy==2.19.0
threadpoolctl==2.2.0
urllib3==1.26.9
Werkzeug==2.0.3
wheel==0.37.1
win-inet-pton==1.1.0
wincertstore==0.2
